Upd: Add more support for builtin function SCCP. TODO: Eliminate parseInt(toString(x)) , print(toString(x)) recursively. Note that toString(parseInt(x)) cannot be eliminated!
What's more, in some cases, the length of an array is a constant.
We may find its definition and replace all call to size with definition.
If its definition is known but not a constant, we may not replace it.
This is out of consideration of the register-allocation pressure.
We will look for all call to size(array).

Finally, inline for builtin function should be supported in ASM.
And a relatively aggressive inliner is needed.

As for loop invariant code motion and loop unrolling, I give up.